+++ nightwatch.conf.js
@@ -1,26 +1,17 @@
 /* eslint import/no-extraneous-dependencies: ["error", {"devDependencies": true}] */
+const nightwatchConfig = require('terra-toolkit/lib/nightwatch/nightwatch.config.js').default;
+const webpackConfig = require('./packages/terra-site/webpack.config');
+const getPackageTestDirectories = require('terra-toolkit/lib/nightwatch/setup-helper.js').getPackageTestDirectories;
 
-const path = require('path');
-const fs = require('fs');
-const loadJsonFile = require('load-json-file');
-const globSync = require('glob').sync;
-const testSettings = require('terra-toolkit/lib/index').testSettings;
-const nightwatchConfiguration = require('terra-toolkit/lib/nightwatch.json');
-const config = require('terra-toolkit/lib/webpack.config');
+let srcFolders;
+const isRepoTest = !process.cwd().includes('/packages/');
 
-module.exports = ((settings) => {
-  const updatedSettings = testSettings(config, settings);
-  updatedSettings.globals_path = './node_modules/terra-toolkit/lib/globals.js';
+if (isRepoTest) {
+  srcFolders = getPackageTestDirectories('lerna.json');
+} else {
+  srcFolders = 'tests/nightwatch/';
+}
 
-  updatedSettings.src_folders = [];
-  loadJsonFile.sync(path.resolve('lerna.json')).packages.forEach((globPath) => {
-    globSync(globPath).forEach((packagePath) => {
-      const nightWatchPath = `${packagePath}/tests/nightwatch`;
-      if (fs.existsSync(nightWatchPath)) {
-        updatedSettings.src_folders.push(nightWatchPath);
-      }
-    });
-  });
+const config = nightwatchConfig(webpackConfig, srcFolders);
 
-  return updatedSettings;
-})(nightwatchConfiguration);
+module.exports = config;