+++ packages/terra-arrange/tests/nightwatch/arrange-spec.js
@@ -1,43 +1,34 @@
-/* eslint import/no-extraneous-dependencies: ["error", {"devDependencies": true}] */
 // eslint-disable-next-line import/no-extraneous-dependencies
-const screenshot = require('terra-toolkit').screenshot;
+const resizeTo = require('terra-toolkit/lib/nightwatch/responsive-helpers').resizeTo;
 
-module.exports = {
-  before: (browser, done) => {
-    browser.resizeWindow(browser.globals.width, browser.globals.height, done);
-  },
-
-  afterEach: (browser, done) => {
-    screenshot(browser, 'terra-arrange', done);
-  },
-
+module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous'], {
   'Displays an arrange with fitStart, fill, and fitEnd content aligned together': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/arrange-tests/aligned-all-arrange`);
+      .url(`${browser.launchUrl}/#/tests/arrange-tests/aligned-all-arrange`);
   },
 
   'Displays an arrange with fill and fitEnd content aligned together': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/arrange-tests/aligned-all-arrange-fit-end-fill`);
+      .url(`${browser.launchUrl}/#/tests/arrange-tests/aligned-all-arrange-fit-end-fill`);
   },
 
   'Displays an arrange with fitStart and fill content aligned together': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/arrange-tests/aligned-all-arrange-fit-start-fill`);
+      .url(`${browser.launchUrl}/#/tests/arrange-tests/aligned-all-arrange-fit-start-fill`);
   },
 
   'Displays an arrange with fitStart aligned individually': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/arrange-tests/aligned-individually-fit-start`);
+      .url(`${browser.launchUrl}/#/tests/arrange-tests/aligned-individually-fit-start`);
   },
 
   'Displays an arrange with fill aligned individually': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/arrange-tests/aligned-individually-fill`);
+      .url(`${browser.launchUrl}/#/tests/arrange-tests/aligned-individually-fill`);
   },
 
   'Displays an arrange with fitEnd aligned individually': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/arrange-tests/aligned-individually-fit-end`);
+      .url(`${browser.launchUrl}/#/tests/arrange-tests/aligned-individually-fit-end`);
   },
-};
+});