+++ packages/terra-form/tests/nightwatch/input/input-spec.js
@@ -1,25 +1,16 @@
-/* eslint import/no-extraneous-dependencies: ["error", {"devDependencies": true}] */
 // eslint-disable-next-line import/no-extraneous-dependencies
-const screenshot = require('terra-toolkit').screenshot;
+const resizeTo = require('terra-toolkit/lib/nightwatch/responsive-helpers').resizeTo;
 
-module.exports = {
-  before: (browser, done) => {
-    browser.resizeWindow(browser.globals.width, browser.globals.height, done);
-  },
-
-  afterEach: (browser, done) => {
-    screenshot(browser, 'terra-form', done);
-  },
-
+module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous'], {
   'Displays a default Input': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/form-tests/input/default`)
+      .url(`${browser.launchUrl}/#/tests/form-tests/input/default`)
       .assert.elementPresent('input');
   },
 
   'Properly initializes Input with the correct attributes': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/form-tests/input/populated`)
+      .url(`${browser.launchUrl}/#/tests/form-tests/input/populated`)
       .assert.elementPresent('input[name="foo"][value="bar"][required]');
   },
-};
+});