+++ packages/terra-modal-manager/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Changed
 * Import breakpoints from CSS modules
+* Update nightwatch tests and test scripts.
 
 1.7.0 - (August 31, 2017)
 ------------------